-class Coordinate added. -Geometry and Color classes extracted from Common.hpp into separate files. -style tag support added
git-svn-id: http://svnrep/sambag/trunk@97 248b3512-769f-48fb-b11e-ddf3fa71ebe4

+++ src/sambag/disco/tests/TestCairoDrawContext.cpp
@@ -405,7 +405,7 @@ void TestCairoDrawContext::testSetStrokeWidth() {
 	IDrawContext::Ptr context = CairoDrawContext::create( surface );
 	//>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>set/compare
 	context->setStrokeWidth(5.0);
-	CPPUNIT_ASSERT_EQUAL ( 5.0, context->getStrokeWidth() );
+	CPPUNIT_ASSERT_EQUAL ( (double)5.0, (double)context->getStrokeWidth() );
 	//
 	context.reset();
 	releaseSurface(surface);
